+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/jdbc/datasource/JDBCBackendDataSourceFactory.java
@@ -33,7 +33,8 @@ public interface JDBCBackendDataSourceFactory {
      *
      * @param dataSourceName data source name
      * @param dataSourceParameter data source connection parameter
+     * @throws Exception when the data source can not be build
      * @return data source for connect backend databases
      */
-    DataSource build(String dataSourceName, DataSourceParameter dataSourceParameter);
+    DataSource build(String dataSourceName, DataSourceParameter dataSourceParameter) throws Exception;
 }
